+++ src/sentry/api/endpoints/organization_spans_fields.py
@@ -16,8 +16,8 @@ from sentry.api.serializers import serialize
 from sentry.api.utils import handle_query_errors
 from sentry.models.organization import Organization
 from sentry.models.project import Project
-from sentry.search.events.builder import SpansIndexedQueryBuilder
 from sentry.search.events.builder.base import BaseQueryBuilder
+from sentry.search.events.builder.spans_indexed import SpansIndexedQueryBuilder
 from sentry.search.events.types import ParamsType, QueryBuilderConfig, SnubaParams
 from sentry.services.hybrid_cloud.organization import RpcOrganization
 from sentry.snuba.dataset import Dataset
