--- src/sentry/search/events/builder/__init__.py
@@ -1,69 +1,0 @@
-from .discover import (
-    HistogramQueryBuilder,
-    TimeseriesQueryBuilder,
-    TopEventsQueryBuilder,
-    UnresolvedQuery,
-)
-from .errors import ErrorsQueryBuilder
-from .issue_platform import IssuePlatformTimeseriesQueryBuilder
-from .metrics import (
-    AlertMetricsQueryBuilder,
-    HistogramMetricQueryBuilder,
-    MetricsQueryBuilder,
-    TimeseriesMetricQueryBuilder,
-    TopMetricsQueryBuilder,
-)
-from .metrics_summaries import MetricsSummariesQueryBuilder
-from .profile_functions import (
-    ProfileFunctionsQueryBuilder,
-    ProfileFunctionsTimeseriesQueryBuilder,
-    ProfileTopFunctionsTimeseriesQueryBuilder,
-)
-from .profile_functions_metrics import (
-    ProfileFunctionsMetricsQueryBuilder,
-    TimeseriesProfileFunctionsMetricsQueryBuilder,
-    TopProfileFunctionsMetricsQueryBuilder,
-)
-from .profiles import ProfilesQueryBuilder, ProfilesTimeseriesQueryBuilder
-from .sessions import SessionsV2QueryBuilder, TimeseriesSessionsV2QueryBuilder
-from .spans_indexed import (
-    SpansIndexedQueryBuilder,
-    TimeseriesSpanIndexedQueryBuilder,
-    TopEventsSpanIndexedQueryBuilder,
-)
-from .spans_metrics import (
-    SpansMetricsQueryBuilder,
-    TimeseriesSpansMetricsQueryBuilder,
-    TopSpansMetricsQueryBuilder,
-)
-
-__all__ = [
-    "HistogramQueryBuilder",
-    "TimeseriesQueryBuilder",
-    "IssuePlatformTimeseriesQueryBuilder",
-    "TopEventsQueryBuilder",
-    "UnresolvedQuery",
-    "AlertMetricsQueryBuilder",
-    "ErrorsQueryBuilder",
-    "HistogramMetricQueryBuilder",
-    "MetricsQueryBuilder",
-    "MetricsSummariesQueryBuilder",
-    "TimeseriesMetricQueryBuilder",
-    "SpansMetricsQueryBuilder",
-    "ProfilesQueryBuilder",
-    "ProfilesTimeseriesQueryBuilder",
-    "ProfileFunctionsQueryBuilder",
-    "ProfileFunctionsTimeseriesQueryBuilder",
-    "ProfileTopFunctionsTimeseriesQueryBuilder",
-    "ProfileFunctionsMetricsQueryBuilder",
-    "TimeseriesProfileFunctionsMetricsQueryBuilder",
-    "TopProfileFunctionsMetricsQueryBuilder",
-    "SessionsV2QueryBuilder",
-    "SpansIndexedQueryBuilder",
-    "TimeseriesSpanIndexedQueryBuilder",
-    "TopEventsSpanIndexedQueryBuilder",
-    "TimeseriesSessionsV2QueryBuilder",
-    "TimeseriesSpansMetricsQueryBuilder",
-    "TopMetricsQueryBuilder",
-    "TopSpansMetricsQueryBuilder",
-]
