+++ src/sentry/search/events/builder/issue_platform.py
@@ -1,5 +1,5 @@
 from sentry.issues.query import manual_group_on_time_aggregation
-from sentry.search.events.builder import TimeseriesQueryBuilder
+from sentry.search.events.builder.discover import TimeseriesQueryBuilder
 from sentry.search.events.types import ParamsType, QueryBuilderConfig, SelectType
 from sentry.snuba.dataset import Dataset
