+++ src/sentry/search/events/builder/profile_functions.py
@@ -5,8 +5,8 @@ from snuba_sdk.function import Function
 
 from sentry.api.event_search import SearchFilter, SearchKey, SearchValue
 from sentry.discover.arithmetic import categorize_columns
-from sentry.search.events.builder import TimeseriesQueryBuilder
 from sentry.search.events.builder.base import BaseQueryBuilder
+from sentry.search.events.builder.discover import TimeseriesQueryBuilder
 from sentry.search.events.datasets.profile_functions import ProfileFunctionsDatasetConfig
 from sentry.search.events.fields import custom_time_processor, get_function_alias
 from sentry.search.events.types import (
